Eine ganze Menge geändert und neu erstellt.
Alles bekomme ich wohl nicht mehr zusammen, hier deswegen nur das
wichtigste.

- abt_transaction_base in abt_transaction umbenannt
- abt_job_ctrl erstellt
- page_log widget erstellt

Übertragung von Aufträgen zur Bank und die rudimentäre Auswertung der
Antworten implementiert.


git-svn-id: https://schmufu.dyndns.org/svn/ab_transfers/trunk@13 459a2fdc-afd4-4fb9-a8a9-e939c3f66bf0

+++ src/widgets/bankaccountswidget.cpp
@@ -104,7 +104,7 @@ BankAccountsWidget::BankAccountsWidget(aqb_Accounts *accounts, QWidget *parent) 
 				ItemCount++;
 				Item->setData(0, Qt::DisplayRole, i.value()->Number());
 				Item->setData(0, Qt::UserRole, i.value()->get_ID());
-				Item->setData(0, Qt::UserRole+1, i.value()->get_AB_ACCOUNT());
+				Item->setData(0, Qt::UserRole+1, (quint64)i.value()->get_AB_ACCOUNT());
 				Item->setData(1, Qt::DisplayRole, i.value()->Name());
 				Item->setData(2, Qt::DisplayRole, i.value()->AccountType());
 				Item->setData(3, Qt::DisplayRole, i.value()->Currency());
